refactor: get and delete team hooks
