+++ frontend/src/components/CreateBoard/SplitBoard/SubTeamsTab/TeamSubTeamsConfigurations.tsx
@@ -4,10 +4,10 @@ import Flex from '@/components/Primitives/Layout/Flex/Flex';
 import Text from '@/components/Primitives/Text/Text';
 import { createBoardError, createBoardTeam } from '@/store/createBoard/atoms/create-board.atom';
 import { TeamUserRoles } from '@/utils/enums/team.user.roles';
-import { useTeams } from '@/hooks/useTeam';
 import { TeamUser } from '@/types/team/team.user';
 import { User } from '@/types/user/user';
 import useCurrentSession from '@/hooks/useCurrentSession';
+import useTeams from '@/hooks/teams/useTeams';
 import MainBoardCard from './MainBoardCard';
 import QuickEditSubTeams from './QuickEditSubTeams';
 import SelectTeam from './SelectTeam';
